@@ -10,7 +10,7 @@ readme = "README.md"
 keywords = ["forecast"]
 license = { text = "MIT" }
 authors = [{ name = "monosans", email = "hsyqixco@protonmail.com" }]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 classifiers = [
   "Framework :: Pydantic",
   "Framework :: Pydantic :: 2",
@@ -19,7 +19,6 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
